pyproject.toml
[tool.poetry]
name = "flask"
version = "1.1.4"
description = ""
authors = ["Yuliia Kopysova <julykopysova97@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.0.1"
psycopg2-binary = "^2.9.1"
Flask-Login = "^0.5.0"
pylint-flask = "^0.6"
pylint-flask-sqlalchemy = "^0.2.0"
flask-marshmallow = "^0.14.0"
marshmallow-sqlalchemy = "^0.26.1"
Flask-RESTful = "^0.3.9"

[tool.poetry.dev-dependencies]
pylint = "2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
